# StreamSyncBackend/docs/openapi.yaml
openapi: 3.0.3
info:
  title: StreamSync API
  version: 1.0.0
  description: Real-time collaborative workspace backend. Workspaces, projects, documents,
    tasks, activity and AI assistance.
paths:
...
  /api/documents/:
    get:
      operationId: documents_list
      description: Bodies are omitted; `excerpt` carries the preview.
      summary: List documents
      parameters:
      - in: query
        name: ordering
        schema:
          type: string
          enum:
          - -created_at
          - -title
          - -updated_at
          - created_at
          - title
          - updated_at
        description: Sort field. Defaults to -updated_at.
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      - in: query
        name: project
        schema:
          type: string
        description: Filter by project id, or "none" for unfiled documents.
      - in: query
        name: search
        schema:
          type: string
        description: Match title or body.
      - in: query
        name: workspace
        schema:
          type: string
        description: Filter by workspace id.
      tags:
      - documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocumentSummaryList'
          description: ''
    post:
      operationId: documents_create
      description: Creates the document and its version 1 in one transaction. Requires
        the editor or owner role in the target workspace.
      summary: Create a document
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentCreateRequest'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentSummary'
          description: ''
        '400':
          description: Project belongs to another workspace.
        '403':
          description: Viewers cannot create documents.
        '404':
          description: No such workspace for this caller.
  /api/documents/{document_id}/:
    get:
      operationId: documents_retrieve
      description: Includes the body and the current revision number.
      summary: Document detail
      parameters:
      - in: path
        name: document_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
          description: ''
        '404':
          description: Not found, or not the caller's workspace.
    patch:
      operationId: documents_update
      description: 'Requires the editor or owner role. Send `revision` to make the
        write conditional: a mismatch returns 409 instead of silently overwriting
        somebody else''s save.'
      summary: Edit a document
      parameters:
      - in: path
        name: document_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedDocumentUpdateRequest'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
          description: ''
        '403':
          description: Viewers cannot edit.
        '409':
          description: The document changed while editing.
    delete:
      operationId: documents_destroy
      description: Requires the editor or owner role. Removes its versions with it.
      summary: Delete a document
      parameters:
      - in: path
        name: document_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      security:
      - jwtAuth: []
      responses:
        '204':
          description: Deleted.
  /api/documents/{document_id}/versions/:
    get:
      operationId: documents_versions_list
      description: Newest first. Snapshot bodies are omitted — restore reads the content
        server-side, so no client needs it.
      summary: List a document's versions
      parameters:
      - in: path
        name: document_id
        schema:
          type: string
          format: uuid
        required: true
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      - name: page_size
        required: false
        in: query
        description: Number of results to return per page.
        schema:
          type: integer
      tags:
      - documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedDocumentVersionList'
          description: ''
  /api/documents/{document_id}/versions/{version_id}/restore/:
    post:
      operationId: documents_version_restore
      description: |-
        Restoring version 5 creates version 6 containing version 5's content. Version 5 is never modified and versions 6..N are never removed — undoing a restore is just another restore.

        Requires the editor or owner role. Returns the updated document.
      summary: Restore a version
      parameters:
      - in: path
        name: document_id
        schema:
          type: string
          format: uuid
        required: true
      - in: path
        name: version_id
        schema:
          type: string
          format: uuid
        required: true
      tags:
      - documents
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentDetail'
          description: ''
        '403':
          description: Viewers cannot restore.
        '404':
          description: No such document or version.
...
  /api/search/:
    get:
      operationId: search_list
      description: Scoped to the caller's workspaces. Results are ranked across types
        and capped; queries shorter than two characters return an empty list.
      summary: Search documents, projects, tasks and people
      parameters:
      - in: query
        name: q
        schema:
          type: string
        description: The query. Under two characters returns [].
      - in: query
        name: workspace
        schema:
          type: string
        description: Optional workspace id to narrow the search to one workspace.
      tags:
      - search
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SearchResult'
          description: ''
...
components:
  schemas:
...
    Collaborator:
      type: object
      description: |-
        The minimal user shape used inside other resources.

        Avatar stacks, "last edited by" labels and member lists need a name and a
        picture, not an email address. Keeping this separate from UserSerializer
        means embedding a user somewhere new does not quietly widen how much
        personal data that endpoint discloses.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        name:
          type: string
          readOnly: true
          title: Full name
          description: Display name shown to collaborators.
        avatar_url:
          type: string
          readOnly: true
      required:
      - avatar_url
      - id
      - name
...
    DocumentCreateRequest:
      type: object
      description: |-
        Creation input.

        `workspace_id` and `project_id` are validated against the caller's own
        access in the view; neither id is trusted as given. The author comes from
        the session, never the body.
      properties:
        workspace_id:
          type: string
          format: uuid
        title:
          type: string
          minLength: 1
          maxLength: 200
        project_id:
          type: string
          format: uuid
          nullable: true
        content:
          type: string
          default: ''
      required:
      - title
      - workspace_id
    DocumentDetail:
      type: object
      description: The full record, including the body the editor loads.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workspace_id:
          type: string
          format: uuid
          readOnly: true
        project_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        project_name:
          type: string
          nullable: true
          readOnly: true
        title:
          type: string
          readOnly: true
        excerpt:
          type: string
          readOnly: true
        author:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        last_edited_by:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        collaborators:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: |-
            Everyone who has contributed to this document.

            Derived from the people on the record — its author, its last editor and
            the authors of its versions — rather than stored on a join table. There
            is no separate notion of "contributor" to maintain, and the set is
            exactly the union of those, so a table would be a second source of
            truth that could disagree.
          readOnly: true
        active_collaborator_ids:
          type: array
          items:
            type: string
          description: |-
            Who is in the document right now.

            Always empty in this milestone. Presence is delivered over the
            WebSocket, which does not exist until Milestone 7 — so nobody is
            connected, and an empty list is the truthful answer rather than a
            placeholder. (README §42)
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        content:
          type: string
          readOnly: true
        revision:
          type: integer
          readOnly: true
      required:
      - active_collaborator_ids
      - author
      - collaborators
      - content
      - created_at
      - excerpt
      - id
      - last_edited_by
      - project_id
      - project_name
      - revision
      - title
      - updated_at
      - workspace_id
...
    DocumentSummary:
      type: object
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        workspace_id:
          type: string
          format: uuid
          readOnly: true
        project_id:
          type: string
          format: uuid
          readOnly: true
          nullable: true
        project_name:
          type: string
          nullable: true
          readOnly: true
        title:
          type: string
          readOnly: true
        excerpt:
          type: string
          readOnly: true
        author:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        last_edited_by:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        collaborators:
          type: array
          items:
            type: object
            additionalProperties: {}
          description: |-
            Everyone who has contributed to this document.

            Derived from the people on the record — its author, its last editor and
            the authors of its versions — rather than stored on a join table. There
            is no separate notion of "contributor" to maintain, and the set is
            exactly the union of those, so a table would be a second source of
            truth that could disagree.
          readOnly: true
        active_collaborator_ids:
          type: array
          items:
            type: string
          description: |-
            Who is in the document right now.

            Always empty in this milestone. Presence is delivered over the
            WebSocket, which does not exist until Milestone 7 — so nobody is
            connected, and an empty list is the truthful answer rather than a
            placeholder. (README §42)
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - active_collaborator_ids
      - author
      - collaborators
      - created_at
      - excerpt
      - id
      - last_edited_by
      - project_id
      - project_name
      - title
      - updated_at
      - workspace_id
    DocumentVersion:
      type: object
      description: |-
        One entry in a document's history.

        The body is deliberately absent: a history list renders "Version 12 / Raj /
        10:42 AM / Added Apple Pay section", and shipping every snapshot's full
        text would make the list heavier than the document itself. Restore reads
        the content server-side, so no client ever needs it.
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        number:
          type: integer
          readOnly: true
        author:
          allOf:
          - $ref: '#/components/schemas/Collaborator'
          readOnly: true
        summary:
          type: string
          readOnly: true
          description: One line describing the change, e.g. "Added Apple Pay section".
        is_current:
          type: boolean
          description: |-
            Whether this is the version the editor is currently showing.

            Compared against a number the view supplies once for the whole page,
            rather than re-querying the maximum for every row.
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - author
      - created_at
      - id
      - is_current
      - number
      - summary
...
    PaginatedDocumentSummaryList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSummary'
    PaginatedDocumentVersionList:
      type: object
      required:
      - count
      - page
      - page_size
      - total_pages
      - results
      properties:
        count:
          type: integer
          example: 42
        page:
          type: integer
          example: 1
        page_size:
          type: integer
          example: 25
        total_pages:
          type: integer
          example: 2
        next:
          type: string
          nullable: true
          format: uri
        previous:
          type: string
          nullable: true
          format: uri
        results:
          type: array
          items:
            $ref: '#/components/schemas/DocumentVersion'
...
    PatchedDocumentUpdateRequest:
      type: object
      description: |-
        Partial update.

        `workspace_id` is absent by design: moving a document between workspaces
        would carry it across a tenant boundary.
      properties:
        title:
          type: string
          minLength: 1
          maxLength: 200
        content:
          type: string
        project_id:
          type: string
          format: uuid
          nullable: true
        revision:
          type: integer
          minimum: 1
        summary:
          type: string
          default: ''
          maxLength: 200
...
    SearchResult:
      type: object
      properties:
        id:
          type: string
          readOnly: true
        type:
          allOf:
          - $ref: '#/components/schemas/SearchResultTypeEnum'
          readOnly: true
        title:
          type: string
          readOnly: true
        subtitle:
          type: string
          readOnly: true
          nullable: true
        href:
          type: string
          readOnly: true
          nullable: true
        score:
          type: number
          format: double
          readOnly: true
      required:
      - href
      - id
      - score
      - subtitle
      - title
      - type
    SearchResultTypeEnum:
      enum:
      - document
      - project
      - task
      - person
      type: string
      description: |-
        * `document` - document
        * `project` - project
        * `task` - task
        * `person` - person
...
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT